requirements.txt
opencv-python>=4.5.0
numpy>=1.19.0
Pillow>=8.0.0
svglib>=1.2.1
reportlab>=3.6.8
PyMuPDF>=1.18.0